+++ CHANGELOG.md
@@ -10,4 +10,6 @@
  - support for no_std
  - support for serde
  - support for rust_decimal
- - support for `From<&str>`, `From<Interval>` and `FromStr` traits
+ - support for `Eq`, `From<&str>`, `From<Interval>` and `FromStr` traits
+ - support for `Hash` trait
+ - support for `Ord`, `PartialOrd` traits